6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 /**
  * main - print sizes of data types
+ * @void: indicate no parameter
  *
  * Return: 0
 */
@@ -13,11 +14,11 @@ int main(void)
 	long long int d;
 	float f;
 
-	printf("size of a char: %d byte(s)\n", (unsigned long)sizeof(a));
-	printf("size of a int: %d byte(s)\n", (unsigned long)sizeof(b));
-	printf("size of a long int: %d byte(s)\n", (unsigned long)sizeof(c));
-	printf("size of a long long int: %d byte(s)\n", (unsigned long)sizeof(d));
-	printf("size of a float: %d byte(s)\n", (unsigned long)sizeof(f));
+	printf("size of a char: %d byte(s)\n", sizeof(a));
+	printf("size of a int: %d byte(s)\n", sizeof(b));
+	printf("size of a long int: %d byte(s)\n", sizeof(c));
+	printf("size of a long long int: %d byte(s)\n", sizeof(d));
+	printf("size of a float: %d byte(s)\n", sizeof(f));
 	return (0);
 
 }
